+++ lib/Exceptions/RedirectUriValidationException.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * SPDX-FileCopyrightText: 2022-2025 Thorsten Jagel <dev@jagel.net>
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ */
+namespace OCA\OIDCIdentityProvider\Exceptions;
+
+class RedirectUriValidationException extends \Exception {
+}